refactor: 음식점 필터/정렬 상태 useRestaurantList에서 관리
